# src/MarqSpec.TradingCopilot.Client/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Trading Co-Pilot API",
    "description": "The single-operator futures trading co-pilot BFF (R-10). Generated from the live minimal-API routes (gh#604) — the source of truth the README links to, replacing the hand-kept table.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate with email + password; returns a JWT bearer token.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "The currently authenticated user.",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
    "/auth/invitations": {
      "post": {
        "tags": ["Auth"],
        "summary": "Issue an invitation to create an account (primary operator only).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInvitationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [{}]
      }
    },
    "/auth/accept-invite": {
      "post": {
        "tags": ["Auth"],
        "summary": "Redeem an invitation token to create an account; returns a JWT.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInviteRequest": {
        "required": ["token", "password", "displayName"],
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The invite token from the invitation."
          },
          "password": {
            "type": "string",
            "description": "The chosen plaintext password."
          },
          "displayName": {
            "type": "string",
            "description": "The chosen display name."
          }
        },
        "description": "A request to accept an invitation and create the account."
      },
...
      "IssueInvitationRequest": {
        "required": ["email"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email to invite."
          }
        },
        "description": "A request to issue an onboarding invitation."
      },
...
      "LoginRequest": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The user's email."
          },
          "password": {
            "type": "string",
            "description": "The user's plaintext password."
          }
        },
        "description": "A login request."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT issued by POST /auth/login, sent as: Authorization: Bearer {token}.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}